map_project/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
psycopg2-binary==2.9.9
psycopg2==2.9.9
django-leaflet==0.26.0
pyproj==3.2.1
django-geojson==4.0.0
gdal==3.4.3
django-cors-headers==3.10.0
pika==1.1.0
